[342] fix Failed to deserialzie UserData when null cycling_ftp_source
