+++ ReleaseNotes.md
@@ -7,6 +7,7 @@
 
 - [#331] Fixed - System.IO.IOException: The process cannot access the file 'deviceInfo.xml' because it is being used by another process.
 - [#289] Fixed - Better logs and messaging when no convert Formats are configured in Settings
+- [#342] Fixed - Failed to deserialzie UserData when `null` cycling_ftp_source
 
 ## Changes
