+++ src/Common/Dto/Peloton/UserData.cs
@@ -9,7 +9,7 @@ public class UserData
 	/// When Source == Ftp_Manual_Source then use Cycling_Ftp
 	/// </summary>
 	[JsonConverter(typeof(JsonStringEnumConverter))]
-	public CyclingFtpSource Cycling_Ftp_Source { get; set; }
+	public CyclingFtpSource? Cycling_Ftp_Source { get; set; }
 	/// <summary>
 	/// Actual FTP generated from a workout
 	/// </summary>
